+++ scratch.js
@@ -1,7 +1,9 @@
 var nlp = require('./src/index');
-// nlp.verbose(true);
+nlp.verbose(true);
 
-var str = `• Spencęr & JOhn™ ⟨lmt⟩.`;
+var str = 'i wrote';
+// var str = 'the FBI';
 var doc = nlp(str);
-
-console.log(doc.out('text'));
+// console.log(doc.values().toNumber());
+// console.log(doc.match(['hello', 'academy']));
+doc.terms().debug();