+++ src/01-doc/selections/people/Person.js
@@ -1,0 +1,13 @@
+//build this class
+const subClass = function(Document) {
+
+  class Person extends Document {
+    nickname() {
+      console.log('it works!');
+      return this;
+    }
+  }
+  return Person;
+
+};
+module.exports = subClass;