+++ src/01-doc/selections/terms/Term.js
@@ -1,0 +1,13 @@
+//build this class
+const subClass = function(Document) {
+
+  class Term extends Document {
+    isTerm() {
+      console.log('it works!');
+      return this;
+    }
+  }
+  return Term;
+
+};
+module.exports = subClass;