+++ src/01-doc/selections/values/Value.js
@@ -1,0 +1,13 @@
+//build this class
+const subClass = function(Document) {
+
+  class Value extends Document {
+    toNumber() {
+      console.log('it works!');
+      return this;
+    }
+  }
+  return Value;
+
+};
+module.exports = subClass;