+++ src/tagger/01-lookups/index.js
@@ -1,7 +1,7 @@
 const checkLexicon = require('./01-lexicon');
-const checkRegex = require('./02-regex');
-const checkPunctuation = require('./03-punctuation');
-const checkEmoji = require('./04-emoji');
+const checkPunctuation = require('./02-punctuation');
+const checkEmoji = require('./03-emoji');
+const checkRegex = require('./04-regex');
 
 //'lookups' look at a term by itself
 const lookups = function(doc) {
@@ -13,16 +13,17 @@ const lookups = function(doc) {
   //try these other methods
   for(let i = 0; i < terms.length; i += 1) {
     let term = terms[i];
-    //don't overwrite existing tags
-    // if (Object.keys(term.tags).length > 0) {
-    //   continue;
-    // }
-    //maybe we can guess
-    checkRegex(term, world);
     //or maybe some helpful punctuation
     checkPunctuation(terms, i, world);
     // ¯\_(ツ)_/¯
     checkEmoji(term, world);
+
+    //don't overwrite existing tags
+    if (Object.keys(term.tags).length > 0) {
+      continue;
+    }
+    //maybe we can guess
+    checkRegex(term, world);
   }
   return doc;
 };