+++ src/world/misc/misc.js
@@ -1,0 +1,17 @@
+//words that can't be compressed, for whatever reason
+module.exports = {
+  '20th century fox': 'Organization',
+  '3m': 'Organization',
+  '7 eleven': 'Organization',
+  '7-eleven': 'Organization',
+  'g8': 'Organization',
+  'motel 6': 'Organization',
+  'vh1': 'Organization',
+  'q1': 'Date',
+  'q2': 'Date',
+  'q3': 'Date',
+  'q4': 'Date',
+  //misc
+  'records': 'Plural',
+  '&': 'Conjunction',
+};